requirements.txt
azure-ai-formrecognizer
azure-core
openai
gradio
python-docx
python-dotenv
requests